Added a popup panel to ask for user choice. Also change the UI color a bit.
The popup panel will appear when the user clicks the camera icon in the home page. The pop will ask the user whether or not they want to take a photo or take a video. The corresponding function will be invoked when the button is clicked. The resulting file is saved to the native gallery app on the mobile device.